pyproject.toml
[tool.poetry]
name = "age-and-gender-recognition"
version = "0.1.0"
description = ""
authors = ["Rowan Mankoo <rowan.mankoo@gmail.com>"]
readme = "README.md"
packages = [{include = "age_and_gender_recognition"}]

[tool.poetry.dependencies]
python = "3.10.0"
torch = "^2.0.0"
numpy = "^1.24.2"
pandas = "^2.0.0"
pillow = "^9.5.0"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
scipy = "^1.10.1"
scikit-learn = "^1.2.2"
torchmetrics = "^0.11.4"
tqdm = "^4.65.0"
mediapipe = "^0.9.3.0"
opencv-python = "^4.7.0.72"
pyyaml = "^6.0"
pydantic = "^1.10.7"
plotly = "^5.15.0"
nbformat = "^5.9.0"
torchvision = "^0.18.0"
pytorch-lightning = "^2.2.4"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
